:sparkles: extract location form
