+++ pages/dashboard/add.vue
@@ -1,33 +1,18 @@
 <script lang="ts" setup>
 import type { FetchError } from "ofetch";
 
-import { toTypedSchema } from "@vee-validate/zod";
-
-import type { NominatimResult } from "~/lib/types";
-
-import { CENTER_JP } from "~/lib/constants";
-import { InsertLocation } from "~/lib/db/schema";
+import type { InsertLocation } from "~/lib/db/schema";
 
 const { $csrfFetch } = useNuxtApp();
-const router = useRouter();
-const mapStore = useMapStore();
 const loading = ref(false);
 const submitted = ref(false);
 const submitError = ref("");
+const submitErrors = ref<Record<string, string>>({});
 
-const { handleSubmit, errors, meta, setErrors, setFieldValue, controlledValues } = useForm({
-  validationSchema: toTypedSchema(InsertLocation),
-  initialValues: {
-    name: "",
-    description: "",
-    long: (CENTER_JP as [number, number])[0],
-    lat: (CENTER_JP as [number, number])[1],
-  },
-});
-
-const onSubmit = handleSubmit(async (values) => {
+async function onSubmit(values: InsertLocation) {
   try {
     submitError.value = "";
+    submitErrors.value = {};
     loading.value = true;
     await $csrfFetch("/api/locations", {
       method: "post",
@@ -39,60 +24,12 @@ const onSubmit = handleSubmit(async (values) => {
   catch (e) {
     const error = e as FetchError;
     if (error.data?.data) {
-      setErrors(error.data?.data);
+      submitErrors.value = error.data?.data;
     }
     submitError.value = getFetchErrorMessage(error);
   }
   loading.value = false;
-});
-
-function formatNumber(value?: number) {
-  if (!value) {
-    return 0;
-  }
-  return value.toFixed(5);
 }
-
-function searchResultSelected(result: NominatimResult) {
-  setFieldValue("name", result.display_name);
-  mapStore.addedPoint = {
-    id: 1,
-    name: "Added Point",
-    description: "",
-    long: Number(result.lon),
-    lat: Number(result.lat),
-    centerMap: true,
-  };
-}
-
-effect(() => {
-  if (mapStore.addedPoint) {
-    setFieldValue("long", mapStore.addedPoint.long);
-    setFieldValue("lat", mapStore.addedPoint.lat);
-  }
-});
-
-onMounted(() => {
-  mapStore.addedPoint = {
-    id: 1,
-    name: "Added point",
-    description: "",
-    long: (CENTER_JP as [number, number])[0],
-    lat: (CENTER_JP as [number, number])[1],
-  };
-});
-
-onBeforeRouteLeave(() => {
-  if (!submitted.value && meta.value.dirty) {
-    // eslint-disable-next-line no-alert
-    const confirm = window.confirm("Are you sure you want to leave? All unsaved changes will be lost.");
-    if (!confirm) {
-      return false;
-    }
-  }
-  mapStore.addedPoint = null;
-  return true;
-});
 </script>
 
 <template>
@@ -115,67 +52,11 @@ onBeforeRouteLeave(() => {
       <Icon name="tabler:circle-letter-x" size="24" />
       <span>{{ submitError }}</span>
     </div>
-
-    <form
-      action=""
-      class="flex flex-col gap-2"
-      @submit.prevent="onSubmit"
-    >
-      <AppFormField
-        name="name"
-        label="Name"
-        :error="errors.name"
-        :disabled="loading"
-      />
-      <AppFormField
-        name="description"
-        label="Description"
-        type="textarea"
-        :error="errors.description"
-        :disabled="loading"
-      />
-      <p class="text-xs text-gray-400">
-        Current coordinates: {{ formatNumber(controlledValues.lat) }}, {{ formatNumber(controlledValues.long) }}
-      </p>
-      <p>To set the coordinates:</p>
-      <ul class="list-disc ml-4 text-sm">
-        <li>
-          Drag the <Icon name="tabler:map-pin-filled" class="text-warning" /> marker on the map.
-        </li>
-        <li>
-          Double click the map.
-        </li>
-        <li>
-          Search for a location below.
-        </li>
-      </ul>
-
-      <div class="flex justify-end gap-2">
-        <button
-          :disabled="loading"
-          type="button"
-          class="btn btn-outline"
-          @click="router.back()"
-        >
-          <Icon name="tabler:arrow-left" size="24" />
-          Cancel
-        </button>
-        <button
-          :disabled="loading"
-          type="submit"
-          class="btn btn-primary"
-        >
-          Add
-          <span v-if="loading" class="loading loading-spinner loading-sm" />
-          <Icon
-            v-else
-            name="tabler:circle-plus-filled"
-            size="24"
-          />
-        </button>
-      </div>
-    </form>
-    <div class="divider" />
-    <AppPlaceSearch @result-selected="searchResultSelected" />
+    <LocationForm
+      :on-submit
+      :loading
+      :submitted
+      :submit-errors
+    />
   </div>
 </template>
